+++ CHANGELOG.rst
@@ -3,7 +3,14 @@ CHANGELOG
 
 This project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+[0.3.0] - 2020-10-26
+  - Adding support for Data propagation report generation
+  - Added 'sig-label' as the new cli option under coverage to capture DP reports
+  - Added support in sail parsers to extract mnemonics also from the trace file
+  - added pytablewriter as part of the requirements
+
 [0.2.0] - 2020-10-23
+
   - Added documentation for CGF and usage
   - Added normalization routine as cli
   - Added abstract functions